feat: :sparkles: Implementation of synchronization with API Rest Spotify to obtain New Realeases, Playlist Features and Categories
Synchronization with API Rest Spotify was implemented to obtain New Realeases, Playlist Features and Categories.  A small refactor was also made to implement architecture by domains, in this case Hexagonal with Patron repository and factory, to ensure that the components are decoupled from the functionalities of the domain and application definitions.